requirement.txt
###### Requirements without Version Specifiers ######
kafka-python
Flask==1.0.2
Flask-SQLAlchemy
pymongo
redis
requests
datetime
boto3
#psycopg2